Updated Dockerfiles in examples to use exec format.
Several examples did not work on Windows due to the CRLF line endings
breaking the entrypoint declarations specified in the Dockerfiles.
This PR both fixes that issue, and brings the Dockerfiles more in line
with recommended Dockerfile best practices by changing the
entrypoint declarations to exec format for the affected examples.

Docker compose version was also standardized.

Signed-off-by: Charles Desbiens <therealcharlesdesbiens@gmail.com>